normalise generation data on-the-fly
